Move get dates functionality to new fetcher
